Pass through all unhandled responses to ChromeCastSpontaneousEventListener
The documentation says that receivers will send spontaneous events to devices w/ the
request id = 0, but in practice this is not true. If I control a chromecast from my
phone I'm receiving updates in this library with requestIds that are absolutely invalid.

For example, this library creates id's that start at 0. On a fresh launch of this app
if I adjust the volume via my phone it is reporting ids in the 200's.